ref: move test_urllib3_timeout to test_snuba
that way the test module name matches
